requirements.txt
# PyTorch -------------------------------------
torch==1.10.1+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
torchvision==0.11.2+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html

# nms_rotated ---------------------------------
-e yolov5_obb/utils/nms_rotated/.

# Base ----------------------------------------
imutils
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.5.4
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Server --------------------------------------
pyzmq
flask

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
thop  # FLOPs computation
